requirements.txt
pyyaml
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=2.0.0
